requirements.txt
Django==2.2.7
psycopg2==2.8.4,<3.0
djangorestframework==3.10.3
ipython==7.9.0
faker==2.0.3
django-cors-headers==3.1.1